pyproject.toml
[project]
name = "url-shorter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.3",
    "pydantic-settings>=2.13.1",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.48",
]
